reorganise les ressources
