--- app/src/main/res/drawable-v23/background_splash.xml
@@ -1,12 +1,0 @@
-<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <item android:drawable="@color/gold"/>
-
-    <item
-        android:drawable="@drawable/max_logo_white"
-        android:gravity="center"
-        android:width="253dp"
-        android:height="70dp"
-        />
-
-</layer-list>
