+++ app/src/main/res/drawable/ic_logo_splash.xml
@@ -1,0 +1,14 @@
+<vector android:height="50dp" android:viewportHeight="102.03"
+    android:viewportWidth="496.54" android:width="243.3304dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#fff" android:pathData="M24.98,6.6q10.23,0 17.48,3.43a25.53,25.53 0,0 1,11.2 9.91,29.57 29.57,0 0,1 4,15.67A36.5,36.5 0,0 1,55.93 46.55,27.63 27.63,0 0,1 50.53,56.26a26.66,26.66 0,0 1,-9.64 7,35.54 35.54,0 0,1 -14.5,2.65L12.53,65.91v34.31L0,100.22L0,6.6ZM26.28,53.6a21.09,21.09 0,0 0,9.19 -1.75,14.74 14.74,0 0,0 5.7,-4.59 19.78,19.78 0,0 0,3 -5.83,19.27 19.27,0 0,0 0.9,-5.57 20.78,20.78 0,0 0,-0.9 -5.89,15.35 15.35,0 0,0 -3,-5.56A16.16,16.16 0,0 0,35.53 20.26a19.83,19.83 0,0 0,-8.48 -1.62L12.53,18.64v35Z"/>
+    <path android:fillColor="#fff" android:pathData="M68.36,0.03L80.53,0.03L80.53,100.21L68.36,100.21Z"/>
+    <path android:fillColor="#fff" android:pathData="M120.66,101.51a31.86,31.86 0,0 1,-15.79 -3.69,26.15 26.15,0 0,1 -10.23,-10.1A29.29,29.29 0,0 1,91.01 73.03a26.42,26.42 0,0 1,4.08 -14.37,30.76 30.76,0 0,1 10.81,-10.36 29.52,29.52 0,0 1,15 -3.88q10.62,0 17.61,6.21t9.58,17.09L103.83,83.25l-2.85,-7.12 36.13,-13.34 -2.46,1.82a17.56,17.56 0,0 0,-5.18 -7.38,14 14,0 0,0 -9.45,-3.11 16.24,16.24 0,0 0,-14.76 8.74,19.71 19.71,0 0,0 -2.21,9.52 20.43,20.43 0,0 0,2.27 9.71,16.61 16.61,0 0,0 6.34,6.66 17.93,17.93 0,0 0,9.26 2.4,18.34 18.34,0 0,0 6.8,-1.3 30.43,30.43 0,0 0,6.28 -3.36l5.44,8.93a36.92,36.92 0,0 1,-9 4.4A30.27,30.27 0,0 1,120.66 101.51Z"/>
+    <path android:fillColor="#fff" android:pathData="M156.66,27.19a6.05,6.05 0,0 1,2.26 -4.79,7.46 7.46,0 0,1 5,-1.94 7.24,7.24 0,0 1,4.92 1.94,6.15 6.15,0 0,1 2.2,4.79A6.09,6.09 0,0 1,168.83 32.03a7.36,7.36 0,0 1,-4.92 1.88,7.59 7.59,0 0,1 -5,-1.88A6,6 0,0 1,156.66 27.19ZM157.43,46.61h12.17v53.6L157.43,100.21Z"/>
+    <path android:fillColor="#fff" android:pathData="M207.15,101.51A28.84,28.84 0,0 1,193.53 98.27a24,24 0,0 1,-9.77 -9.52q-3.63,-6.27 -3.63,-15.34a30.52,30.52 0,0 1,3.75 -15.54,26.11 26.11,0 0,1 10.1,-10 28.15,28.15 0,0 1,13.86 -3.5A22.78,22.78 0,0 1,220.53 47.84a23.15,23.15 0,0 1,7.83 8.09l-0.77,1.94 1.16,-11.26h11.14v53.6h-12v-14l1.42,3.11a16.35,16.35 0,0 1,-2.46 3.3,31.35 31.35,0 0,1 -4.6,4.08 24.71,24.71 0,0 1,-15.08 4.79ZM210.39,91.51A19,19 0,0 0,218.53 89.85,15.8 15.8,0 0,0 224.53,85.03a18,18 0,0 0,3.31 -7.25L227.84,67.33a16.53,16.53 0,0 0,-3.63 -6.8A18.18,18.18 0,0 0,218.03 56.03a19.15,19.15 0,0 0,-7.9 -1.62,17.54 17.54,0 0,0 -8.94,2.33 17.36,17.36 0,0 0,-6.4 6.47,19.05 19.05,0 0,0 -2.4,9.71 18.62,18.62 0,0 0,2.46 9.46,18.37 18.37,0 0,0 6.6,6.73A17.22,17.22 0,0 0,210.39 91.54Z"/>
+    <path android:fillColor="#fff" android:pathData="M276.94,101.51a25.31,25.31 0,0 1,-23 -13.66,31.1 31.1,0 0,1 -3.56,-15.09 29.4,29.4 0,0 1,3.76 -15.27,25.34 25.34,0 0,1 10.1,-9.71A29.82,29.82 0,0 1,278.53 44.41a28.11,28.11 0,0 1,12.11 2.72,21.74 21.74,0 0,1 9,7.51l-0.64,4.27L299,0.03h12L311,100.21h-12L299,86.88l1.42,3.36a30.67,30.67 0,0 1,-9.9 8.1A28.67,28.67 0,0 1,276.94 101.51ZM280.69,91.28a19,19 0,0 0,8.48 -1.81,18.75 18.75,0 0,0 6.22,-4.86 17.9,17.9 0,0 0,3.56 -7.18L298.95,67.2a16.64,16.64 0,0 0,-3.69 -6.87,18.26 18.26,0 0,0 -6.35,-4.59 20.23,20.23 0,0 0,-8.22 -1.62,18 18,0 0,0 -8.93,2.33 17.54,17.54 0,0 0,-6.61 6.47,19.77 19.77,0 0,0 0,19.1 18.86,18.86 0,0 0,6.61 6.73A16.8,16.8 0,0 0,280.69 91.28Z"/>
+    <path android:fillColor="#fff" android:pathData="M351.25,101.51a31.86,31.86 0,0 1,-15.79 -3.69,26.08 26.08,0 0,1 -10.23,-10.1A29.29,29.29 0,0 1,321.6 73.03a26.42,26.42 0,0 1,4.08 -14.37A30.76,30.76 0,0 1,336.53 48.29a29.52,29.52 0,0 1,15 -3.88q10.62,0 17.61,6.21t9.58,17.09L334.42,83.25 331.53,76.13l36.13,-13.34 -2.46,1.82a17.56,17.56 0,0 0,-5.18 -7.38,14 14,0 0,0 -9.45,-3.11 16.24,16.24 0,0 0,-14.76 8.74,19.71 19.71,0 0,0 -2.21,9.52 20.43,20.43 0,0 0,2.27 9.71,16.61 16.61,0 0,0 6.34,6.66 17.93,17.93 0,0 0,9.26 2.4,18.34 18.34,0 0,0 6.8,-1.3 30.43,30.43 0,0 0,6.28 -3.36l5.44,8.93a36.92,36.92 0,0 1,-9 4.4A30.27,30.27 0,0 1,351.25 101.51Z"/>
+    <path android:fillColor="#fff" android:pathData="M411.91,14.26h23.74v72.34h-23.74z"/>
+    <path android:fillColor="#fff" android:pathData="M439.49,4.98h23.74v81.39h-23.74z"/>
+    <path android:fillColor="#fff" android:pathData="M466.59,21.31h23.74v65.06h-23.74z"/>
+    <path android:fillColor="#fff" android:pathData="M406.19,90.16h90.35v11.87h-90.35z"/>
+</vector>
