--- app/src/main/res/font/josefin.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<font-family xmlns:android="http://schemas.android.com/apk/res/android">
-    <font
-        android:font="@font/josefin_sans_regular"
-        android:fontStyle="normal"
-        android:fontWeight="400" />
-</font-family>
