--- app/src/main/res/values-v21/styles.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-
-    <style name="SplashTheme" parent="Theme.AppCompat.NoActionBar">
-        <item name="colorPrimaryDark">@color/gold</item>
-        <item name="android:windowBackground">@drawable/background_splash</item>
-        <item name="android:windowAnimationStyle">@null</item>
-        <item name="android:navigationBarColor">@color/gold</item>
-    </style>
-</resources>
