+++ app/src/main/res/values/styles.xml
@@ -12,4 +12,11 @@
         <item name="android:windowAnimationStyle">@null</item>
     </style>
 
+    <style name="text_md">
+        <item name="android:fontFamily">sans-serif-light</item>
+        <item name="fontStyle">normal</item>
+        <item name="fontWeight">400</item>
+        <item name="android:textSize">18sp</item>
+    </style>
+
 </resources>
